--- a/cel/options.go
+++ b/cel/options.go
@@ -219,7 +219,7 @@ func Functions(funcs ...*functions.Overload) ProgramOption {
 func Globals(vars interface{}) ProgramOption {
 	return func(p *prog) (*prog, error) {
 		defaultVars, err :=
-			interpreter.NewAdaptingActivation(p.adapter, vars)
+			interpreter.NewActivation(vars)
 		if err != nil {
 			return nil, err
 		}