--- a/cel/options.go
+++ b/cel/options.go
@@ -229,6 +229,15 @@ func Globals(vars interface{}) ProgramOption {
 }
 
 // CustomResolver adjusts the variable and field resolution behavior of the Program.
+//
+// An expression containing a field references and/or indexing expressions is considered an
+// Attribute. An Attribute is a variable or object reference with zero or more Qualifier that
+// values determine which piece of data is relevant to the computation. When an Attribute is
+// encountered, the runtime uses a Resolver to determine how to map from the Attribute to a
+// concrete value.
+//
+// A Resolver may be generic or it may be deeply aware of the objects which can appear within
+// CEL expressions. By default a generic Resolver is created if a custom one is not set.
 func CustomResolver(resolver interpreter.Resolver) ProgramOption {
 	return func(p *prog) (*prog, error) {
 		p.resolver = resolver