--- a/interpreter/BUILD.bazel
+++ b/interpreter/BUILD.bazel
@@ -45,7 +45,6 @@ go_test(
         "attributes_test.go",
         "interpreter_test.go",
         "prune_test.go",
-        "resolver_test.go",
     ],
     embed = [
         ":go_default_library",