There are three high-level concepts introduced in this PR:

- `Attribute`: a variable or object reference whose use may be qualified zero or more field / index 
   accesses.
- `Qualifier`: a valid protobuf map key type (`bool`, `int`, `string`, `uint`) used to select data from
   a value.
- `Resolver`: a mechanism for determining how to resolve an `Attribute` to a value from an 
  `Activation` or some other context in the expression under evaluation.

The use of the `Attribute` and `Resolver` changes the default evaluation semantics for missing data by producing an error unless expressly instructed to produce an `types.Unknown` value (#190). The performance is also greatly improved for many data accesses.

Before:
```
select_key-8         	  100000	     11079 ns/op	    2433 B/op	     118 allocs/op
select_index-8       	  200000	     11930 ns/op	    2769 B/op	     121 allocs/op
select_field-8       	  100000	     10445 ns/op	    2801 B/op	     103 allocs/op
select_pb2_primitive_fields-8         	  100000	     13927 ns/op	    2513 B/op	     103 allocs/op
select_pb3_wrapper_fields-8           	  200000	     10586 ns/op	    2245 B/op	      77 allocs/op
select_relative-8                     	 1000000	      1665 ns/op	     744 B/op	      17 allocs/op
```

After:
```
select_key-8 	           1000000	      1543 ns/op	     160 B/op	      18 allocs/op
select_index-8             1000000	      1874 ns/op	     416 B/op	      22 allocs/op
select_field-8             2000000	       986 ns/op	     208 B/op	       9 allocs/op
select_pb2_primitive_fields-8         	  500000	      3456 ns/op	     464 B/op	      23 allocs/op
select_pb3_wrapper_fields-8           	 1000000	      1485 ns/op	     404 B/op	       5 allocs/op
select_relative-8                     	 1000000	      1484 ns/op	     728 B/op	      16 allocs/op
```

The After results also come with changes in the test harness which will no longer pre-allocate inputs as `ref.Val`, so the improvements will likely be greater than for most than what's stated here.

Here are two additional points of comparison for single protobuf field comparisons. The first is with the default resolver, the second uses a custom resolver:

```
select_pb3_compare-8          10000000	       176 ns/op	      16 B/op	       2 allocs/op
select_custom_pb3_compare-8   10000000	       135 ns/op	      16 B/op	       2 allocs/op
```